Add files via upload
version TWO makes the record txt file on the local machine become unreadable, by using base64 and url encode with some string replacement. Yet, some bugs are need to be improve.
V2 decrypter helps the txt file become readble, so that the analyze.py are able to to create a accessable table for analyzing use.